refactor: 디렉터리 구조 리팩터링
