+++ src/js/App.js
@@ -2,7 +2,7 @@ import { headerEventHandler } from './Header/headerEventHandler.js';
 import { headerView } from './Header/headerView.js';
 import { inputBarEventHandler } from './InputBar/inputBarEventHandler.js';
 import { inputBarView } from './InputBar/inputBarView.js';
-import { monthlyHistoryView } from './MainPage/monthlyHistoryView.js';
+import { monthlyHistoryView } from './MainPage/View/monthlyHistoryView.js';
 import { storage } from './Storage.js';
 
 export function App() {
