+++ src/js/MainPage/View/dailyHistoryView.js
@@ -1,5 +1,5 @@
-import { CLASSLIST, TAG_NAME, WEEKDAY } from '../constant.js';
-import { createNode, formatMoney } from '../utils.js';
+import { CLASSLIST, TAG_NAME, WEEKDAY } from '../../constant.js';
+import { createNode, formatMoney } from '../../utils.js';
 import { dailyItemView } from './dailyItemView.js';
 
 export const dailyHistoryView = (dailyHistory) => {
