+++ src/js/MainPage/View/dailyItemView.js
@@ -1,5 +1,5 @@
-import { CLASSLIST, FILE_PATH, MONEY_TYPE, TAG_NAME } from '../constant.js';
-import { createNode, formatMoney } from '../utils.js';
+import { CLASSLIST, FILE_PATH, MONEY_TYPE, TAG_NAME } from '../../constant.js';
+import { createNode, formatMoney } from '../../utils.js';
 
 export const dailyItemView = (dailyItem) => {
   const { uuid, category, memo, payment, money, isIncomeMoney } = dailyItem;
